+++ src/Section.java
@@ -1,0 +1,7 @@
+/**
+ * Enum class for the section (watchlist, currWatch, or watched) of
+ * AnimeLibrary.
+ */
+public enum Section {
+    Watchlist, currWatch, Watched
+}