+++ src/Tier.java
@@ -1,0 +1,6 @@
+/**
+ * Enum class for tiers of shows.
+ */
+public enum Tier {
+    SS, S, A, B, C, D, E, Unranked
+}